refactor: rename route files and update import paths for UserController
